requirements.txt
pandas-datareader==0.10.0
matplotlib==3.0.3
requests==2.26.0
pymongo==3.12.0
certifi==2021.5.30
scikit-image==0.14.2
scikit-learn==0.23.0
scipy==1.2.1
seaborn==0.9.0
selenium==3.141.0
calplot==0.1.7.3
statsmodels==0.11.0
pytorch==1.7.0